refactor(base): remove unused interface implementation
BaseInterface and its family interfaces require Activity context but
they does not have any context in its own. So they acquire a context
from the top activity using the static methods of ActivityReference.
However this is not necessary if a view model does not handle any
operations related to View components. Furthermore, this method could
lead to unexpected behaviors if a view model acquires the wrong context
because it always gets a context from the top in the stack of activities.

The main changes of this commit is that there will be no more global
contexts except for the Application context. Any operations that need
context must be done in Activity or Fragment, or just use the application
context. For example, navigating to another fragment, which is
apparently a view related operation, must be done in view code. This
change also removes dependency to View from view model thus making test
more applicable.