refactor(base): remove redundant BaseRecyclerAdapter
We already have an alternative in androidx. Also BaseRecyclerAdapter
forces a child class to have one layout binding, this may cause a problem
when it comes to flexibility. And current implementation of
DiffUtilCallback compares an object as a whole, which will make us to
loose performance.

This commit removes BaseRecyclerAdapter and replaces it with ListAdapter
in androidx. Furthermore, it adds SimpleDiffUtil to reduce code
duplication but gives flexibility to choose what property to compare
with when calculating differences of entries.