refactor: remove copy/pasted code
OpenKLAS does not permit copy/pasting others' work of any kind as copy
and pasting may cause potential license issues. Even if one did not
violate the license, copy/pasting is not a polite way to respect the
others' hard work.

RWInterceptor does not seem to be a copied code but it is not actually
needed in the project, I removed the whole class and injection code.