fix: request is interrupted on configuration change
There is no reason for view model to inherit LifecycleObserver. The only
usage of it is to hook onDestroy() to clear CompositeDisposable, but
this only results in stopping ongoing request on configuration change
which is not a desired behavior. ViewModel of AAC has its own lifecycle,
so we just need to clear CompositeDisposable at onClear() phase.

This commit removes inheritance of LifecycleObserver from ViewModel
which is handling lifecycle wrongly. Also I have removed inheritance of
Observer from ViewModel because it seems there is no need to observe
ViewModel as a whole.