fix(session): arbitrary error should not navigate user to login
Checks if our request gave valid session by checking if the server
orders us to redirect.

This can be done without KlasService using Response data type if we
create our own CallAdapter, but this is error-prone job and affects
badly in terms of maintainability.

The code assumes that the redirection implies session is invalid, we
need further observations to confirm that this assumption is correct.